OpenAI_Agents_SDK/deep_research_pro/pyproject.toml
[project]
name = "deep-research-pro"
version = "0.1.0"
description = "Deep Research Pro (MVP) using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai-agents>=0.0.17",
  "openai>=1.85.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.0.0",
  "httpx==0.27.2",
  "certifi>=2024.0.0",
  "rich>=13.7.0",
  "gradio>=4.0.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
drp = "app.run:main"